Redirects to tables after submitting forms
After submitting add*.php and edit*.php forms, the site no longer
display a success message with link back to the tables. Instead, it uses
js to redirect users to the tables, with a success message displayed
above the tables which users can click to close.

Fix problem of how disabling startDay and endDay options in income forms
causes error in submitting forms. Forms no longer dictates that startDay
and endDay of incomes must be the same for every month.